move team code from engine to teams
