+++ go/teams/get_test.go
@@ -1,6 +1,7 @@
 package teams
 
 import (
+	"encoding/hex"
 	"testing"
 	"time"
 
@@ -101,7 +102,12 @@ func teamGet(t *testing.T) {
 }
 
 func createTeam(tc libkb.TestContext) string {
-	name, err := kbtest.CreateTeam(tc.G)
+	b, err := libkb.RandBytes(4)
+	if err != nil {
+		tc.T.Fatal(err)
+	}
+	name := hex.EncodeToString(b)
+	err = CreateRootTeam(context.TODO(), tc.G, name)
 	if err != nil {
 		tc.T.Fatal(err)
 	}
